4115A - Formulary Collection Process - additional clean-up items
More adjustments to the CSS hardcoded widths.  Why the fsck does
Chrome change the width on the tier columns?!
Tried setting the classname for the first *AND* last rows -
thinking that the rendering engine wouldn't muck with the td
element width after recognizing the CSS class at the table's
last row.

This means war.  Next up: What if we get rid of the colspan attribute ...